remove ITraceable
